Moved NavigationBar component with all its elements to new folder "core"
