+++ src/core/NavigationBar/SearchBar/index.js
