+++ src/core/NavigationBar/SearchBar/styled.js
