+++ src/core/NavigationBar/styled.js
